setup/requirements_setup.txt
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2